feat: move GeoHash, Hashtag, and Relay feed filters + ViewModels to commons
Moves the feed filter logic and ViewModel base classes for GeoHash, Hashtag,
and Relay feeds into commons module using IAccount + ICacheProvider abstractions.

Android-side retains thin wrappers that:
- Provide LocalCache as the ICacheProvider
- Include ViewModelProvider.Factory for AndroidX integration

Pattern for remaining ~27 FeedViewModel subclasses to follow.
Part of KMP iOS migration tracked in #2238.
