+++ amethyst/src/main/java/com/vitorpamplona/amethyst/ui/screen/loggedIn/relay/dal/RelayFeedFilter.kt
@@ -20,37 +20,5 @@
  */
 package com.vitorpamplona.amethyst.ui.screen.loggedIn.relay.dal
 
-import com.vitorpamplona.amethyst.model.Account
-import com.vitorpamplona.amethyst.model.LocalCache
-import com.vitorpamplona.amethyst.model.Note
-import com.vitorpamplona.amethyst.ui.dal.AdditiveFeedFilter
-import com.vitorpamplona.amethyst.ui.dal.DefaultFeedOrder
-import com.vitorpamplona.quartz.nip01Core.relay.normalizer.NormalizedRelayUrl
-
-class RelayFeedFilter(
-    val relayUrl: NormalizedRelayUrl,
-    val account: Account,
-    val cache: LocalCache,
-) : AdditiveFeedFilter<Note>() {
-    override fun feedKey(): String = account.userProfile().pubkeyHex + "-" + relayUrl.url
-
-    override fun feed(): List<Note> {
-        val notes =
-            cache.notes.filterIntoSet { _, it ->
-                acceptableEvent(it)
-            }
-
-        return sort(notes)
-    }
-
-    override fun applyFilter(newItems: Set<Note>): Set<Note> = innerApplyFilter(newItems)
-
-    private fun innerApplyFilter(collection: Collection<Note>): Set<Note> = collection.filterTo(HashSet()) { acceptableEvent(it) }
-
-    fun acceptableEvent(it: Note): Boolean =
-        it.hasRelay(relayUrl) &&
-            !it.isHiddenFor(account.hiddenUsers.flow.value) &&
-            account.isAcceptable(it)
-
-    override fun sort(items: Set<Note>): List<Note> = items.sortedWith(DefaultFeedOrder)
-}
+// Re-export from commons for backwards compatibility
+typealias RelayFeedFilter = com.vitorpamplona.amethyst.commons.viewmodels.relay.RelayFeedFilter
