+++ commons/src/commonMain/kotlin/com/vitorpamplona/amethyst/commons/viewmodels/relay/RelayFeedViewModel.kt
@@ -18,7 +18,20 @@
  * AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package com.vitorpamplona.amethyst.ui.screen.loggedIn.geohash.dal
+package com.vitorpamplona.amethyst.commons.viewmodels.relay
 
-// Re-export from commons for backwards compatibility
-typealias GeoHashFeedFilter = com.vitorpamplona.amethyst.commons.viewmodels.geohash.GeoHashFeedFilter
+import androidx.compose.runtime.Stable
+import com.vitorpamplona.amethyst.commons.model.IAccount
+import com.vitorpamplona.amethyst.commons.model.cache.ICacheProvider
+import com.vitorpamplona.amethyst.commons.viewmodels.FeedViewModel
+import com.vitorpamplona.quartz.nip01Core.relay.normalizer.NormalizedRelayUrl
+
+@Stable
+open class RelayFeedViewModel(
+    val relayUrl: NormalizedRelayUrl,
+    val account: IAccount,
+    cacheProvider: ICacheProvider,
+) : FeedViewModel(
+        RelayFeedFilter(relayUrl, account, cacheProvider),
+        cacheProvider,
+    )
